unit test fixes
